+++ src/main/java/racingcar/RacingGameApplication.java
@@ -1,5 +1,9 @@
 package racingcar;
 
+import racingcar.domain.RacingGame;
+import racingcar.view.InputView;
+import racingcar.view.ResultView;
+
 public class RacingGameApplication {
     public static void main(String[] args) {
         InputView inputView = new InputView();
