+++ src/main/java/racingcar/domain/Car.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class Car {
     private int location;
