+++ src/main/java/racingcar/util/StateUtil.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 public class StateUtil {
     private final static String STATE = "-";
