+++ src/test/java/racingcar/CarTest.java
@@ -2,9 +2,9 @@ package racingcar;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racingcar.domain.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
