+++ src/test/java/racingcar/RacingGameTest.java
@@ -2,9 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.ArrayList;
-import java.util.List;
+import racingcar.domain.RacingGame;
 
 import static org.assertj.core.api.Assertions.assertThat;
